backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
motor==3.3.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2
pydantic==2.6.3
email-validator==2.1.0.post1
python-dotenv==1.0.1 